% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transformationGrob.r
\name{transformationGrob}
\alias{transformationGrob}
\title{Affine transformation grob}
\usage{
transformationGrob(
  grob,
  vp.define = NULL,
  transform = NULL,
  name = NULL,
  gp = gpar(),
  vp = NULL
)
}
\arguments{
\item{grob}{A grid grob to perform affine transformations on.}

\item{vp.define}{\code{\link[grid:viewport]{grid::viewport()}} to define grid group in.}

\item{transform}{An affine transformation function.
If \code{NULL} default to \code{\link[grid:viewportTransform]{grid::viewportTransform()}}.}

\item{name}{A character identifier (for grid)}

\item{gp}{A \code{\link[grid:gpar]{grid::gpar()}} object.}

\item{vp}{A \code{\link[grid:viewport]{grid::viewport()}} object (or \code{NULL}).}
}
\description{
\code{transformationGrob()} is a grid grob function to facilitate
using the group affine transformation features introduced in R 4.2.
}
\examples{
if (getRversion() >= '4.2.0' && require("grid")) {
  # Only works if active graphics device supports affine transformations
  # such as `png(type="cairo")` on R 4.2+ but not `ragg::agg_png()`
  grob <- grobTree(circleGrob(gp=gpar(fill="yellow", col="blue")),
                   textGrob("RSTATS", gp=gpar(fontsize=32)))

  vp.define <- viewport(width=unit(2, "in"), height=unit(2, "in"))
  transformation <- transformationGrob(grob, vp.define=vp.define)
  grid.newpage()
  pushViewport(viewport(width=unit(4, "in"), height=unit(1, "in")))
  grid.draw(transformation)
  popViewport()
}

print(pkgdown::fig_settings())
print(names(dev.cur()))
print(isTRUE(dev.capabilities()$transformations))
print(sessionInfo())

}
